Fix GCC -Wall -Wextra -Wpedantic -Werror build on Linux
- Add GCC diagnostic pragmas to suppress intentional Wolf3D patterns:
  -Wswitch (partial enum switches), -Wimplicit-fallthrough, -Wabsolute-value,
  -Wmissing-field-initializers, -Wchar-subscripts, -Wsign-compare,
  -Wformat-overflow, -Wcast-function-type, -Wdiscarded-qualifiers,
  -Wunused-result, -Wunused-variable, -Wunused-function
- Add Linux stubs for MSVC-only functions: itoa/ltoa/ultoa in wl_inter.c,
  wl_text.c, wl_agent.c; wolf_stricmp->strcasecmp in id_us_1.c
- Fix fread() unused-result in id_ca.c with local variable pattern
- Fix &(cast) non-lvalue error in wl_menu.c and wl_play.c with temp var
- Fix menuitems enum in wl_menu.h: add typedef to prevent ODR violation
- Remove duplicate loadedgame definition from wl_main.c (defined in id_us_1.c)
- Strip 0x1A DOS EOF byte from foreign.h to fix GCC "stray \32" error

+++ src/wl_draw.c
@@ -1117,7 +1117,7 @@ void DrawPlayerWeapon (void)
 	}
 #endif
 
-	if (gamestate.weapon != -1)
+	if ((int)gamestate.weapon != -1)
 	{
 		shapenum = weaponscale[gamestate.weapon]+gamestate.weaponframe;
 		SimpleScaleShape(viewwidth/2,shapenum,viewheight+1);
